+++ server/src/jobs/offrePartenaire/clever-connect/atlas/processAtlas.ts
@@ -1,0 +1,6 @@
+import { importAtlasRaw, importAtlasToComputed } from "@/jobs/offrePartenaire/clever-connect/atlas/importAtlas"
+
+export const processAtlas = async () => {
+  await importAtlasRaw()
+  await importAtlasToComputed()
+}
