+++ server/src/jobs/offrePartenaire/clever-connect/meteojob/importMeteojob.test.ts
@@ -4,12 +4,13 @@ import { JOBPARTNERS_LABEL } from "shared/models/jobsPartners.model"
 import { beforeEach, describe, expect, it, vi } from "vitest"
 
 import { getDbCollection } from "@/common/utils/mongodbUtils"
-import { importAtlasRaw, importAtlasToComputed } from "@/jobs/offrePartenaire/veritone/atlas/importAtlas"
 import { useMongo } from "@tests/utils/mongo.test.utils"
+
+import { importMeteojobRaw, importMeteojobToComputed } from "./importMeteojob"
 
 const now = new Date("2024-07-21T04:49:06.000+02:00")
 
-describe("importAtlas", () => {
+describe("importMeteojob", () => {
   useMongo()
 
   beforeEach(() => {
@@ -19,19 +20,19 @@ describe("importAtlas", () => {
     return async () => {
       vi.useRealTimers()
       await getDbCollection("computed_jobs_partners").deleteMany({})
-      await getDbCollection("raw_atlas").deleteMany({})
+      await getDbCollection("raw_meteojob").deleteMany({})
     }
   })
 
-  it("should test the import of Atlas data into computed_job_partners", async () => {
-    const fileStream = fs.createReadStream("server/src/jobs/offrePartenaire/veritone/atlas/importAtlas.test.input.xml")
-    await importAtlasRaw(fileStream)
-    expect.soft(await getDbCollection("raw_atlas").countDocuments({})).toBe(4)
+  it("should test the import of Meteojob data into computed_job_partners", async () => {
+    const fileStream = fs.createReadStream("server/src/jobs/offrePartenaire/clever-connect/meteojob/importMeteojob.test.input.xml")
+    await importMeteojobRaw(fileStream)
+    expect.soft(await getDbCollection("raw_meteojob").countDocuments({})).toBe(3)
 
-    await importAtlasToComputed()
+    await importMeteojobToComputed()
     const jobs = (
       await getDbCollection("computed_jobs_partners")
-        .find({ partner_label: JOBPARTNERS_LABEL.ATLAS }, { projection: { _id: 0, created_at: 0 } })
+        .find({ partner_label: JOBPARTNERS_LABEL.METEOJOB }, { projection: { _id: 0, created_at: 0 } })
         .toArray()
     ).sort((a, b) => ((a.partner_job_id ?? "") < (b.partner_job_id ?? "") ? -1 : 1))
     expect.soft(jobs.length).toBe(3)
