+++ server/src/jobs/offrePartenaire/clever-connect/meteojob/processMeteojob.ts
@@ -1,0 +1,6 @@
+import { importMeteojobRaw, importMeteojobToComputed } from "@/jobs/offrePartenaire/clever-connect/meteojob/importMeteojob"
+
+export const processMeteojob = async () => {
+  await importMeteojobRaw()
+  await importMeteojobToComputed()
+}
