--- server/src/jobs/offrePartenaire/veritone/atlas/processAtlas.ts
@@ -1,6 +1,0 @@
-import { importAtlasRaw, importAtlasToComputed } from "@/jobs/offrePartenaire/veritone/atlas/importAtlas"
-
-export const processAtlas = async () => {
-  await importAtlasRaw()
-  await importAtlasToComputed()
-}
