--- server/src/jobs/offrePartenaire/veritone/meteojob/processMeteojob.ts
@@ -1,6 +1,0 @@
-import { importMeteojobRaw, importMeteojobToComputed } from "./importMeteojob"
-
-export const processMeteojob = async () => {
-  await importMeteojobRaw()
-  await importMeteojobToComputed()
-}
